Move BlockRecorder to proper file
